IMP Install Script is now able to update Vimified.
I add conditions on existing folders in the current one. This allows us to run the install script more than once and do not get fatal error on existing folders.
I also add a trick to avoid the Vim warning.

NOTE: I invite people to beautify the display rendering. It's way ugly but I never did this so if anyone has some times and knowlegde, go on! ,)